made postOrder redux
