+++ src/store/cart/actions.js
@@ -37,39 +37,3 @@ export const addToCart = (id) => {
     }
   };
 };
-
-export const postOrder = (productId) => {
-  return async (dispatch, getState) => {
-    const user = selectUser(getState());
-    const token = selectToken(getState());
-
-    const userId = parseInt(user.id);
-
-    const products = productId;
-    dispatch(appLoading());
-
-    const response = await axios.post(
-      `${apiUrl}/products/order`,
-      {
-        userId,
-        productId,
-      },
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    console.log("Response?", response.data);
-
-    dispatch(setMessage("success", true, response.data.message));
-
-    // dispatch({
-    //   type: "POST_ORDER",
-    //   payload: response.data,
-    // });
-
-    dispatch(appDoneLoading());
-  };
-};
