+++ src/store/cart/reducer.js
@@ -8,9 +8,6 @@ export default (state = initialState, action) => {
 
       return [...state, { ...action.payload }];
 
-    case "POST_ORDER":
-      return action.payload;
-
     default:
       return state;
   }
