+++ src/store/postOrder/reducer.js
@@ -1,0 +1,11 @@
+const initialState = [];
+
+export default (state = initialState, action) => {
+  switch (action.type) {
+    case "POST_ORDER":
+      [...state, { ...action.payload }];
+
+    default:
+      return state;
+  }
+};
