+++ src/store/postOrder/selector.js
@@ -1,0 +1,1 @@
+export const selectOrder = (reduxState) => reduxState.selectOrder;
